% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbank.R
\name{appendSpeechRate}
\alias{appendSpeechRate}
\title{appendSpeechRate: append pre-generated speech rate data to given dataframe t}
\usage{
appendSpeechRate(t, file = NULL)
}
\arguments{
\item{t}{a table read through read.cha}

\item{file}{speech rate file}
}
\value{
dataframe t bound to speech rates per utterance
}
\description{
appendSpeechRate: append pre-generated speech rate data (see audioproc.R)
}
\author{
luzs
}
