% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocalgraphs.R
\docType{data}
\name{atddia}
\alias{atddia}
\title{A sample Medical Team Meeting dialogue encoded as a vocaldia}
\format{A data frame with 7869 rows and 7 variables:
\describe{
  \item{id}{The dialogue indentifier}
  \item{begin}{The start time of a speech turn or silence interval}
  \item{end}{The end time of a speech turn or silence interval}
  \item{speaker}{An identifier for the speaker of the turn, or Floor for silence.}
  \item{role}{The speaker's role (patient, interviewer, other, or Floor}
  \item{trans}{The transcription of the turn (blanked out for anonymity)}
  \item{dx}{The diagnosis (ad or nonad}
 }}
\source{
This dataset was generated from the Carolina Conversations
    Collection, and used in the work described in De La Fuente,
    Albert and Luz:
    "Detecting cognitive decline through dialogue processing",
    2017. For the full data set, please contact the Medical
    University of South Carolina (MUSC)
    http://carolinaconversations.musc.edu/
}
\usage{
atddia
}
\description{
A dataset containing 38 dialogues (17 control patients, and 21 AD
patients) and 7869 vocalisation events.
}
\keyword{datasets}
