% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{getEntropy}
\alias{getEntropy}
\title{getEntropy: safely return the Shannon entropy of a distribution.}
\usage{
getEntropy(distribution)
}
\arguments{
\item{distribution}{a probability distribution.}
}
\value{
a numeric value.
}
\description{
Compute the entropy of a distribution.
}
\details{
Compute the entropy of a distribution.
}
