% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbank.R
\name{getIDs}
\alias{getIDs}
\title{getIDs get speaker role IDs (PAR, INV) and info from CHA content}
\usage{
getIDs(text)
}
\arguments{
\item{text}{a string vector containing the lines of a CHA file}
}
\value{
a vector with participants IDs
}
\description{
getIDs get speaker IDs from CHA content
}
\author{
luzs
}
