% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbank.R
\name{getPID}
\alias{getPID}
\title{getIDs get  study-wide unique patient IDs from CHA content}
\usage{
getPID(text)
}
\arguments{
\item{text}{a string vector containing the lines of a CHA file}
}
\value{
a vector with participants IDs
}
\description{
getPIDs get study-wide unique patient IDs from CHA content
}
\author{
luzs
}
