% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{getPofAgivenB}
\alias{getPofAgivenB}
\title{getPofAgivenB: transtion probability.}
\usage{
getPofAgivenB(a, b, ttarray)
}
\arguments{
\item{a}{target node}

\item{b}{source node}

\item{ttarray}{adjacency matrix}
}
\value{
a transition probability.
}
\description{
Conditional (transition ) probability
}
\details{
Retrieve \eqn{p(a|b)}, probability of a transition from b to a in an
adjacency matrix
}
