% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbank.R
\name{getSilences}
\alias{getSilences}
\title{getSilences read silences file}
\usage{
getSilences(file, sildir = NULL, silsuffix = "c.mp3.csv")
}
\arguments{
\item{file}{CSV formatted silences file}

\item{sildir}{dir where silence files are}

\item{silsuffix}{## suffix for silence files}
}
\value{
silences dataframe
}
\description{
getSilences read silences file
}
\author{
luzs
}
