% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbank.R
\name{getSyllablesAndSilences}
\alias{getSyllablesAndSilences}
\title{getSyllablesAndSilences: process Praat's grid for syllable nuclei}
\usage{
getSyllablesAndSilences(txtgrid)
}
\arguments{
\item{txtgrid}{Path to Praat grid file generated by praat-syllable-syllable-nuclei-v2}
}
\value{
list of syllables and silences
}
\description{
getSyllablesAndSilences: process Praat's grid for syllable nuclei, based on De Jong's approach
}
\references{
De Jong, N. H. and Wempe, T. (2009). Praat script to detect syllable nuclei
 and measure speech rate automatically. Behavior Research Methods,
 41(2):385–390, May.
}
\author{
luzs
}
