% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbank.R
\name{getTranscript}
\alias{getTranscript}
\title{getTranscript: get transcription lines from .cha content}
\usage{
getTranscript(text)
}
\arguments{
\item{text}{a string vector containing the lines of a CHA file}
}
\value{
a list of transcriptions (participant and interviewer utterances)
}
\description{
getTranscript
}
\author{
luzs
}
