% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{identifyGrpVocalisations}
\alias{identifyGrpVocalisations}
\title{identifyGrpVocalisations: replace appropriate vocalisation
    types}
\usage{
identifyGrpVocalisations(vocvector)
}
\arguments{
\item{vocvector}{a character vector containing a sequence of
vocalisation events}
}
\value{
A vector with all events replaced by the appropriate type
    identifier.
}
\description{
Identify group vocalisations
}
\details{
Standardise identifier for group vocalisations
}
\examples{
data(vocdia)
identifyGrpVocalisations(atddia$speaker[1:60])
}
