% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{identifyPauses}
\alias{identifyPauses}
\title{identifyPauses: label pauses according to type.}
\usage{
identifyPauses(vocvector)
}
\arguments{
\item{vocvector}{a character vector containing a sequence of
vocalisation events}
}
\value{
A vector with all Floor events replaced by the appropriate
    pause type identifier.
}
\description{
Assign types to the pauses (Floor events) in a sequence
}
\details{
Identify the pauses in a vector as one of the pauses in
\code{pauseTypes}
}
\examples{
data(vocdia)
identifyPauses(atddia$speaker[1:60])
}
