% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{igraph.vocaldia}
\alias{igraph.vocaldia}
\title{igraph.vocaldia: Create an igraph vocalisation diagram}
\usage{
igraph.vocaldia(vd, ...)
}
\arguments{
\item{vd}{a vocalisation diagram}

\item{...}{arguments for the layout algorithm}
}
\value{
an igraph
}
\description{
Create an igraph vocalisation diagram
}
\details{
Create a vocalisation diagram
}
\examples{
data(vocdia)
if (require('igraph'))
    igraph.vocaldia(getSampledVocalMatrix(subset(atddia,
                                          id=='Abbott_Maddock_01'),
                  individual=TRUE, nodecolumn='speaker'))
}
