% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbank.R
\name{makeSessionDataSet}
\alias{makeSessionDataSet}
\title{makeSessionDataSet: create a data frame for a session (e.g. cookie scene description) based on .cha transcription files}
\usage{
makeSessionDataSet(
  f,
  sildir = NULL,
  silsuffix = "c.mp3.csv",
  srdir = "../data/ADReSS/speech_rate/",
  srsuffix = "sra",
  sprate = T
)
}
\arguments{
\item{f}{CHA file to read}

\item{sildir}{directory where silence profiles are stored}

\item{silsuffix}{suffix for silence files}

\item{srdir}{directory where speech rate csv (1 value per utterance) files are stored}

\item{srsuffix}{the suffix of the speech rate files (default: sre)}

\item{sprate}{estimate speech rate? (default: TRUE)}
}
\value{
a speech session data frame
}
\description{
makeSessionDataSet: create a data frame for a session (e.g. cookie scene description)
}
\author{
luzs
}
