% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{matrixExp}
\alias{matrixExp}
\title{matrixExp: raise matrix to exp.}
\usage{
matrixExp(matrix, exp, mmatrix = matrix)
}
\arguments{
\item{matrix}{a matrix}

\item{exp}{the power to which matrix will be raised}

\item{mmatrix}{a placeholder.}
}
\value{
matrix^exp
}
\description{
Matrix exponentials
}
\details{
A (sort of) exponential function for matrix multiplication (to be
used with \code{\link{staticMatrix}}).
}
\examples{
data(vocdia)
matrixExp(vocmatrix$ttarray, 3)
}
