% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{plot.matrixseries}
\alias{plot.matrixseries}
\title{plotConvergence: plots Markov diagram convergence.}
\usage{
\method{plot}{matrixseries}(x, ..., par = list(), interact = F)
}
\arguments{
\item{x}{an object of class matrixseries; a list where the
\eqn{i^{th}} element corresponds to \eqn{M^i}.}

\item{...}{extra graphics parameters for plot.}

\item{par}{graphic parameters alist}

\item{interact}{if TRUE, pauses the drawing after each node.}
}
\value{
the matrixseries
}
\description{
Visualise  convergence properties of vocalisation graphs
}
\details{
A 'toy' for visualisation of convergence properties of
vocalisation graphs. Plot the convergence paths of each
Vocalisation event (i.e. each row-column transition probability,
grouped by colour according to the inciding node)
}
\examples{
data(vocdia)
plot(staticMatrix(vocmatrix$ttarray, digits=4, history=TRUE))
}
