% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{plot.vocaldia}
\alias{plot.vocaldia}
\title{plot.vocaldia}
\usage{
\method{plot}{vocaldia}(x, ...)
}
\arguments{
\item{x}{a vocalisation diagram}

\item{...}{arguments for the layout algorithm}
}
\value{
\code{NULL}
}
\description{
Plot a vocalisation diagram
}
\details{
Plot a vocalisation diagram
}
\examples{
data(vocdia)
if (require('igraph'))
 plot(getSampledVocalMatrix(subset(atddia, id=='Abbott_Maddock_01'),
                          individual=TRUE, nodecolumn='speaker'))
}
