% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{startmatrix}
\alias{startmatrix}
\alias{startmatrix.default}
\alias{startmatrix.matrixseries}
\title{startmatrix: return the first matrix of a converging series.}
\usage{
startmatrix(mseries)

\method{startmatrix}{default}(mseries)

\method{startmatrix}{matrixseries}(mseries)
}
\arguments{
\item{mseries}{a matrixseries object}
}
\value{
the initial matrix.
}
\description{
Access initital matrix in a \code{matrixseries}
}
\details{
Access initital matrix in a \code{matrixseries}
}
\examples{
data(vocdia)
x2 <- staticMatrix(vocmatrix$ttarray, digits=4, history=TRUE)
## original matrix
 startmatrix(x2)
}
