% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocalgraphs.R
\docType{data}
\name{vocmatrix}
\alias{vocmatrix}
\title{A sample vocalisation matrix}
\format{A list containing 2 arrayes
\describe{
  \item{ttarray}{The vocaldia adjacency matrix}
  \item{tdarray}{The proportional durations (stationary probabilities) of each event (node)}
}}
\source{
This dataset was generated from the Multomodal Leatrning
    Analytics dataset, for the eponymous ICMI'13 Grand
    Challenge. The use these vocaldias were put to is described in
    Luz (2013). The full dataset and code is available
    at https://gitlab.scss.tcd.ie/saturnino.luz/icmi-mla-challenge
}
\usage{
vocmatrix
}
\description{
A \code{vocaldia} object containing a 3-speaker dialogue
}
\references{
S. Luz. Automatic identification of experts and performance
  prediction in the multimodal math data corpus through analysis
  of speech interaction. In \emph{Proceedings of the 15th ACM on
  International Conference on Multimodal Interaction, ICMI'13},
  pages 575--582, New York, NY, USA, 2013. ACM.
}
\keyword{datasets}
