% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{write.vocaldia}
\alias{write.vocaldia}
\title{write.vocaldia}
\usage{
write.vocaldia(vd, file = "", ...)
}
\arguments{
\item{vd}{a vocalisation diagram}

\item{file}{name of file to which dot diagram will be written.}

\item{...}{arguments passed on to toDotNotation. If "", write to STDOUT.}
}
\value{
\code{NULL}
}
\description{
Write vocalisation diagram to file in dot (graphviz) notation
}
\details{
Write a vocalisation diagram
}
\examples{
data(vocdia)
write.vocaldia(getSampledVocalMatrix(subset(atddia, id=='Abbott_Maddock_01'),
                            individual=TRUE, nodecolumn='speaker'),
                            file='/tmp/vocdia.dot')
}
