% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{plot.vocaldia}
\alias{plot.vocaldia}
\title{plot.vocaldia}
\usage{
\method{plot}{vocaldia}(x, ..., package = "igraph")
}
\arguments{
\item{x}{a vocalisation diagram}

\item{...}{arguments for the layout algorithm}

\item{package}{the package to be used for ploting (igraph
(default) or Rgraphviz)}
}
\value{
\code{NULL}
}
\description{
Plot a vocalisation diagram
}
\details{
Plot a vocalisation diagram
}
\examples{
data(vocdia)
require('igraph')
plot(getSampledVocalMatrix(subset(atddia, id=='Abbott_Maddock_01'),
                          individual=TRUE, nodecolumn='speaker'))
}
